backend/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
celery
redis
requests
boto3
langchain
langchain-community
langchain-aws
langgraph
python-dotenv
flask-cors
tenacity
